lsp: move shared lsp files inside lsp directory
