+++ src/dev/flang/lsp/feature/CodeLenses.java
@@ -37,11 +37,11 @@ import org.eclipse.lsp4j.Command;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
 
+import dev.flang.lsp.shared.FeatureTool;
+import dev.flang.lsp.shared.QueryAST;
+import dev.flang.lsp.shared.Util;
 import dev.flang.lsp.util.Bridge;
 import dev.flang.lsp.util.LSP4jUtils;
-import dev.flang.shared.FeatureTool;
-import dev.flang.shared.QueryAST;
-import dev.flang.shared.Util;
 
 public class CodeLenses
 {
