+++ src/dev/flang/lsp/feature/Completion.java
@@ -40,14 +40,14 @@ import org.eclipse.lsp4j.InsertTextMode;
 
 import dev.flang.ast.AbstractFeature;
 import dev.flang.ast.AbstractType;
+import dev.flang.lsp.shared.FeatureTool;
+import dev.flang.lsp.shared.LexerTool;
+import dev.flang.lsp.shared.ParserTool;
+import dev.flang.lsp.shared.QueryAST;
+import dev.flang.lsp.shared.TypeTool;
+import dev.flang.lsp.shared.Util;
 import dev.flang.lsp.util.Bridge;
 import dev.flang.parser.Lexer.Token;
-import dev.flang.shared.FeatureTool;
-import dev.flang.shared.LexerTool;
-import dev.flang.shared.ParserTool;
-import dev.flang.shared.QueryAST;
-import dev.flang.shared.TypeTool;
-import dev.flang.shared.Util;
 
 /**
  * tries offering code completions
