+++ src/dev/flang/lsp/feature/Diagnostics.java
@@ -42,14 +42,14 @@ import org.eclipse.lsp4j.Range;
 
 import dev.flang.ast.AbstractFeature;
 import dev.flang.lsp.Config;
+import dev.flang.lsp.shared.ASTWalker;
+import dev.flang.lsp.shared.FeatureTool;
+import dev.flang.lsp.shared.LexerTool;
+import dev.flang.lsp.shared.ParserTool;
+import dev.flang.lsp.shared.QueryAST;
+import dev.flang.lsp.shared.Util;
 import dev.flang.lsp.util.Bridge;
 import dev.flang.lsp.util.LSP4jUtils;
-import dev.flang.shared.ASTWalker;
-import dev.flang.shared.FeatureTool;
-import dev.flang.shared.LexerTool;
-import dev.flang.shared.ParserTool;
-import dev.flang.shared.QueryAST;
-import dev.flang.shared.Util;
 
 /**
  * provide diagnostics for a given uri
