+++ src/dev/flang/lsp/feature/InlayHints.java
@@ -44,15 +44,15 @@ import dev.flang.ast.AbstractCall;
 import dev.flang.ast.Constant;
 import dev.flang.ast.AbstractFeature;
 import dev.flang.ast.Expr;
+import dev.flang.lsp.shared.ASTWalker;
+import dev.flang.lsp.shared.CallTool;
+import dev.flang.lsp.shared.FeatureTool;
+import dev.flang.lsp.shared.LexerTool;
+import dev.flang.lsp.shared.TypeTool;
+import dev.flang.lsp.shared.Util;
 import dev.flang.lsp.util.Bridge;
 import dev.flang.lsp.util.LSP4jUtils;
 import dev.flang.parser.Lexer.Token;
-import dev.flang.shared.ASTWalker;
-import dev.flang.shared.CallTool;
-import dev.flang.shared.FeatureTool;
-import dev.flang.shared.LexerTool;
-import dev.flang.shared.TypeTool;
-import dev.flang.shared.Util;
 import dev.flang.util.ANY;
 import dev.flang.util.SourcePosition;
