+++ src/dev/flang/lsp/shared/ASTWalker.java
@@ -24,7 +24,7 @@ Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
  *
  *---------------------------------------------------------------------*/
 
-package dev.flang.shared;
+package dev.flang.lsp.shared;
 
 import java.net.URI;
 import java.util.AbstractMap.SimpleEntry;
