+++ src/dev/flang/lsp/shared/ErrorHandling.java
@@ -24,7 +24,7 @@ Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
  *
  *---------------------------------------------------------------------*/
 
-package dev.flang.shared;
+package dev.flang.lsp.shared;
 
 import java.util.Arrays;
 import java.util.concurrent.Callable;
