+++ src/dev/flang/lsp/shared/IO.java
@@ -24,7 +24,7 @@ Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
  *
  *---------------------------------------------------------------------*/
 
-package dev.flang.shared;
+package dev.flang.lsp.shared;
 
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
