+++ src/dev/flang/lsp/shared/SourcePositionTool.java
@@ -24,7 +24,7 @@ Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
  *
  *---------------------------------------------------------------------*/
 
-package dev.flang.shared;
+package dev.flang.lsp.shared;
 
 import dev.flang.util.ANY;
 import dev.flang.util.SourceFile;
